+++ polimorphism 3_2_2/src/ru/netology/taxtypes/TaxType.java
@@ -1,8 +1,10 @@
 package ru.netology.taxtypes;
 
+import java.math.BigDecimal;
+
 public class TaxType {
-    public double calculateTaxFor(double amount) {
-        return 0.0;
+    public BigDecimal calculateTaxFor(BigDecimal amount) {
+        return BigDecimal.valueOf(0.0);
     }
 }
